requirements.txt
django==2.2.3
python-decouple==3.1
mysqlclient==1.4.2
pytest==5.0.1
mysqlclient==1.4.2
django-redis==4.10.0
djangorestframework==3.10.1
django-filter==2.2.0
Markdown==3.1.1